+++ python/ray/serve/drivers.py
@@ -1,17 +1,13 @@
 import functools
-import inspect
 import logging
-from abc import abstractmethod
-from typing import Any, Callable, Optional, Type, Union, Dict
-from pydantic import BaseModel
+from typing import Any, Callable, Optional, Union, Dict
 import ray
 from ray.serve._private.utils import install_serve_encoders_to_fastapi
-from ray.util.annotations import DeveloperAPI, PublicAPI
+from ray.util.annotations import PublicAPI
 
 import starlette
-from fastapi import Body, Depends, FastAPI
+from fastapi import Depends, FastAPI
 
-from ray._private.utils import import_attr
 from ray.serve.deployment_graph import RayServeDAGHandle
 from ray.serve._private.http_util import ASGIHTTPSender
 from ray.serve.handle import RayServeDeploymentHandle
@@ -22,75 +18,9 @@ import asyncio
 import grpc
 from ray.serve.generated import serve_pb2, serve_pb2_grpc
 from ray.serve._private.constants import DEFAULT_GRPC_PORT, SERVE_LOGGER_NAME
-
-DEFAULT_HTTP_ADAPTER = "ray.serve.http_adapters.starlette_request"
-HTTPAdapterFn = Callable[[Any], Any]
+from ray.serve.drivers_utils import load_http_adapter
 
 logger = logging.getLogger(SERVE_LOGGER_NAME)
-
-
-def _load_http_adapter(
-    http_adapter: Optional[Union[str, HTTPAdapterFn, Type[BaseModel]]]
-) -> HTTPAdapterFn:
-    if http_adapter is None:
-        http_adapter = DEFAULT_HTTP_ADAPTER
-
-    if isinstance(http_adapter, str):
-        http_adapter = import_attr(http_adapter)
-
-    if inspect.isclass(http_adapter) and issubclass(http_adapter, BaseModel):
-
-        def http_adapter(inp: http_adapter = Body(...)):
-            return inp
-
-    if not inspect.isfunction(http_adapter):
-        raise ValueError(
-            "input schema must be a callable function or pydantic model class."
-        )
-
-    if any(
-        param.annotation == inspect.Parameter.empty
-        for param in inspect.signature(http_adapter).parameters.values()
-    ):
-        raise ValueError("input schema function's signature should be type annotated.")
-    return http_adapter
-
-
-@DeveloperAPI
-class SimpleSchemaIngress:
-    def __init__(
-        self, http_adapter: Optional[Union[str, HTTPAdapterFn, Type[BaseModel]]] = None
-    ):
-        """Create a FastAPI endpoint annotated with http_adapter dependency.
-
-        Args:
-            http_adapter(str, HTTPAdapterFn, None, Type[pydantic.BaseModel]):
-              The FastAPI input conversion function or a pydantic model class.
-              By default, Serve will directly pass in the request object
-              starlette.requests.Request. You can pass in any FastAPI dependency
-              resolver. When you pass in a string, Serve will import it.
-              Please refer to Serve HTTP adatper documentation to learn more.
-        """
-        install_serve_encoders_to_fastapi()
-        http_adapter = _load_http_adapter(http_adapter)
-        self.app = FastAPI()
-
-        @self.app.get("/")
-        @self.app.post("/")
-        async def handle_request(inp=Depends(http_adapter)):
-            resp = await self.predict(inp)
-            return resp
-
-    @abstractmethod
-    async def predict(self, inp):
-        raise NotImplementedError()
-
-    async def __call__(self, request: starlette.requests.Request):
-        # NOTE(simon): This is now duplicated from ASGIAppWrapper because we need to
-        # generate FastAPI on the fly, we should find a way to unify the two.
-        sender = ASGIHTTPSender()
-        await self.app(request.scope, receive=request.receive, send=sender)
-        return sender.build_asgi_response()
 
 
 @PublicAPI(stability="beta")
@@ -113,7 +43,7 @@ class DAGDriver:
                 HTTP requests to Ray Serve input.
         """
         install_serve_encoders_to_fastapi()
-        http_adapter = _load_http_adapter(http_adapter)
+        http_adapter = load_http_adapter(http_adapter)
         self.app = FastAPI()
 
         if isinstance(dags, dict):
@@ -222,40 +152,29 @@ class gRPCIngress:
         await self.server.wait_for_termination()
 
 
-if not isinstance(serve_pb2_grpc.PredictAPIsServiceServicer, type):
-    # The reason we have this fake Driver class is that sphinx-build is to mock
-    # PredictAPIsServiceServicer and it will cause the metaclass conflict issue
-    # from multiple inheritance.
-    @serve.deployment
-    class DefaultgRPCDriver:
-        def __init__(self):
-            raise Exception("gRPC driver Placeholder, it is not expected to be used")
+@serve.deployment(is_driver_deployment=True, ray_actor_options={"num_cpus": 0})
+class DefaultgRPCDriver(gRPCIngress, serve_pb2_grpc.PredictAPIsServiceServicer):
+    """
+    gRPC Driver that responsible for redirecting the gRPC requests
+    and hold dag handle
+    """
 
-else:
+    def __init__(self, dags: RayServeDAGHandle, port=DEFAULT_GRPC_PORT):
+        """Create a grpc driver based on the PredictAPIsService schema.
 
-    @serve.deployment(is_driver_deployment=True, ray_actor_options={"num_cpus": 0})
-    class DefaultgRPCDriver(gRPCIngress, serve_pb2_grpc.PredictAPIsServiceServicer):
+        Args:
+            dags: a handle to a Ray Serve DAG or a dictionary of handles.
+            port: Port to use to listen to receive the request
         """
-        gRPC Driver that responsible for redirecting the gRPC requests
-        and hold dag handle
+        self.dag = dags
+        # TODO(Sihan) we will add a gRPCOption class
+        # once we have more options to use
+        gRPCIngress.__init__(self, port)
+
+    async def Predict(self, request, context):
         """
+        gRPC Predict function implementation
+        """
+        res = await (await self.dag.remote(dict(request.input)))
 
-        def __init__(self, dags: RayServeDAGHandle, port=DEFAULT_GRPC_PORT):
-            """Create a grpc driver based on the PredictAPIsService schema.
-
-            Args:
-                dags: a handle to a Ray Serve DAG or a dictionary of handles.
-                port: Port to use to listen to receive the request
-            """
-            self.dag = dags
-            # TODO(Sihan) we will add a gRPCOption class
-            # once we have more options to use
-            gRPCIngress.__init__(self, port)
-
-        async def Predict(self, request, context):
-            """
-            gRPC Predict function implementation
-            """
-            res = await (await self.dag.remote(dict(request.input)))
-
-            return serve_pb2.PredictResponse(prediction=res)
+        return serve_pb2.PredictResponse(prediction=res)
